refactor: split gRPC services into modules
Splits the everything-grpc-in-one-file into smaller, per-service
modules.
